requirements.txt
# Decision Governance Layer - Dependencies
# Python 3.11+

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# HTTP client (used by BedrockClient)
httpx==0.25.2

# Utilities
python-dotenv==1.0.0

# Database (auth feature)
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Auth
bcrypt>=4.0.0
PyJWT>=2.8.0
email-validator>=2.0.0

# AWS Bedrock Nova
